requirements.txt
requests==2.17.3
numpy==1.14.3
pandas==0.20.3
scipy==0.19.0
flask-swagger==0.2.13
gunicorn
tensorflow
sklearn
